Team

+++ src/main/java/com/app/todoapp/entity/Team.java
@@ -6,8 +6,8 @@ import java.time.LocalDateTime;
 
 @Data
 public class Team {
-    private int id;
+    private Long id;
     private String name;
     private LocalDateTime createTime;
-    private int createdBy;
+    private Long createdBy;
 }
